--- src/modules/texts/constants/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./text.constants";
-export * from "./text.constants";
